no dependency on db_schema
