+++ elfpy/hyperdrive_interface/hyperdrive_interface.py
@@ -17,9 +17,6 @@ from web3.types import BlockData
 
 from elfpy import eth
 from elfpy import time as elftime
-
-# TODO: We should not be importing db_schema in this file
-from elfpy.data.db_schema import PoolInfo, Transaction, WalletInfo
 from elfpy.markets.hyperdrive import HyperdriveMarket, HyperdriveMarketState, HyperdrivePricingModel, hyperdrive_assets
 
 from .hyperdrive_addresses import HyperdriveAddresses
@@ -105,18 +102,9 @@ def get_hyperdrive_pool_info(web3: Web3, hyperdrive_contract: Contract, block_nu
     current_block_timestamp = current_block.get("timestamp")
     if current_block_timestamp is None:
         raise AssertionError("Current block has no timestamp")
-    pool_info_data_dict.update({"timestamp": current_block_timestamp})
+    pool_info_data_dict.update({"timestamp": datetime.fromtimestamp(current_block_timestamp)})
     pool_info_data_dict.update({"blockNumber": block_number})
     pool_info_dict = {}
-    for key in PoolInfo.__annotations__.keys():
-        # Required keys
-        if key == "timestamp":
-            pool_info_dict[key] = datetime.fromtimestamp(pool_info_data_dict[key])
-        elif key == "blockNumber":
-            pool_info_dict[key] = pool_info_data_dict[key]
-        # Otherwise default to None if not exist
-        else:
-            pool_info_dict[key] = pool_info_data_dict.get(key, None)
     position_duration = eth.smart_contract_read(hyperdrive_contract, "getPoolConfig").get("positionDuration", None)
     if position_duration is not None:
         asset_id = hyperdrive_assets.encode_asset_id(
@@ -221,103 +209,3 @@ def get_hyperdrive_market(web3: Web3, hyperdrive_contract: Contract) -> Hyperdri
             _step_size=FixedPoint(1) / FixedPoint(365),
         ),
     )
-
-
-def get_wallet_info(
-    hyperdrive_contract: Contract,
-    base_contract: Contract,
-    block_number: BlockNumber,
-    transactions: list[Transaction],
-    poolinfo: PoolInfo,
-) -> list[WalletInfo]:
-    """Retrieves wallet information at a given block given a transaction
-    Transactions are needed here to get
-    (1) the wallet address of a transaction, and
-    (2) the token id of the transaction
-
-    Arguments
-    ----------
-    hyperdrive_contract : Contract
-        The deployed hyperdrive contract instance.
-    base_contract : Contract
-        The deployed base contract instance
-    block_number : BlockNumber
-        The block number to query
-    transactions : list[Transaction]
-        The list of transactions to get events from
-
-    Returns
-    -------
-    list[WalletInfo]
-        The list of WalletInfo objects ready to be inserted into postgres
-    """
-    # pylint: disable=too-many-locals
-    out_wallet_info = []
-    for transaction in transactions:
-        wallet_addr = transaction.event_operator
-        token_id = transaction.event_id
-        token_prefix = transaction.event_prefix
-        token_maturity_time = transaction.event_maturity_time
-        if wallet_addr is None:
-            continue
-        # Query and add base tokens to walletinfo
-        num_base_token_scaled = None
-        for _ in range(RETRY_COUNT):
-            try:
-                num_base_token_scaled = base_contract.functions.balanceOf(wallet_addr).call(
-                    block_identifier=block_number
-                )
-                break
-            except ValueError:
-                logging.warning("Error in getting base token balance, retrying")
-                time.sleep(1)
-                continue
-        num_base_token = eth.convert_scaled_value(num_base_token_scaled)
-        if (num_base_token is not None) and (wallet_addr is not None):
-            out_wallet_info.append(
-                WalletInfo(
-                    blockNumber=block_number,
-                    walletAddress=wallet_addr,
-                    baseTokenType="BASE",
-                    tokenType="BASE",
-                    tokenValue=num_base_token,
-                )
-            )
-        # Query and add hyperdrive tokens to walletinfo
-        if (token_id is not None) and (token_prefix is not None):
-            base_token_type = hyperdrive_assets.AssetIdPrefix(token_prefix).name
-            if (token_maturity_time is not None) and (token_maturity_time > 0):
-                token_type = base_token_type + "-" + str(token_maturity_time)
-                maturity_time = token_maturity_time
-            else:
-                token_type = base_token_type
-                maturity_time = None
-            num_custom_token_scaled = None
-            for _ in range(RETRY_COUNT):
-                try:
-                    num_custom_token_scaled = hyperdrive_contract.functions.balanceOf(int(token_id), wallet_addr).call(
-                        block_identifier=block_number
-                    )
-                except ValueError:
-                    logging.warning("Error in getting custom token balance, retrying")
-                    time.sleep(1)
-                    continue
-            num_custom_token = eth.convert_scaled_value(num_custom_token_scaled)
-            if num_custom_token is not None:
-                # Check here if token is short
-                # If so, add share price from pool info to data
-                share_price = None
-                if (base_token_type) == "SHORT":
-                    share_price = poolinfo.sharePrice
-                out_wallet_info.append(
-                    WalletInfo(
-                        blockNumber=block_number,
-                        walletAddress=wallet_addr,
-                        baseTokenType=base_token_type,
-                        tokenType=token_type,
-                        tokenValue=num_custom_token,
-                        maturityTime=maturity_time,
-                        sharePrice=share_price,
-                    )
-                )
-    return out_wallet_info
